backend/requirements.txt
absl-py==1.3.0
ai21==1.2.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
apikey==0.2.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
astor==0.8.1
asttokens==2.1.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
autoscraper==1.1.14
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.5
bot-studio==1.4.0
branca==0.6.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
cloudscraper==1.2.71
codecov==2.1.12
colorama==0.4.6
commonmark==0.9.1
constantly==15.1.0
contourpy==1.0.6
coverage==7.1.0
cryptography==40.0.2
cssselect==1.2.0
cx-Oracle==8.3.0
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.6
dnspython==2.4.1
docutils==0.20.1
entrypoints==0.4
env==0.1.0
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.8.0
Flask==2.2.2
flatbuffers==23.5.8
folium==0.14.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
geographiclib==2.0
geopy==2.3.0
Ghost.py==0.2.3
gingerit==0.9.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.1.0a0
gradio==3.20.1
greenlet==2.0.1
grpcio==1.50.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==0.9999999
httpcore==0.9.1
httpx==0.13.3
hub-sdk==0.0.3
huggingface-hub==0.13.2
hyperframe==5.2.0
hyperlink==21.0.0
identify==2.5.18
idna==3.7
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
imutils==0.5.4
incremental==22.10.0
iniconfig==2.0.0
ipykernel==6.17.0
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jax==0.4.9
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.21.0
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.2
keras==2.6.0
keras-nightly==2.14.0.dev2023051107
keyring==24.2.0
kiwisolver==1.4.4
libclang==16.0.0
linkify-it-py==2.0.0
livereload==2.6.3
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.4
ml-dtypes==0.1.0
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.2.1
multidict==6.0.4
mypy==1.0.0
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.9
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.8.1
nodeenv==1.7.0
notebook==6.5.2
notebook_shim==0.2.2
numpy==2.0.0
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
orjson==3.8.7
packaging==21.3
pafy==0.5.5
pandas==1.5.1
pandocfilters==1.5.0
parsel==1.8.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pkginfo==1.9.6
platformdirs==2.5.3
plotly==5.12.0
pluggy==1.0.0
pre-commit==3.0.4
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.32
Protego==0.2.1
protobuf==3.20.0
psutil==5.9.3
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.2
pydeck==0.8.0
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.4.1
Pympler==1.0.1
pynecone-io==0.1.13
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyQRCode==1.2.1
pyrsistent==0.19.2
pyshorteners==1.0.1
pytesseract==0.3.10
pytest==7.2.1
pytest-cov==4.0.0
python-barcode==0.14.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
python-multipart==0.0.6
pytube==15.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==305
pywin32-ctypes==0.2.2
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qrcode==7.4.2
qtconsole==5.4.0
QtPy==2.3.0
queuelib==1.6.2
readme-renderer==40.0
redis==4.4.2
regex==2022.10.31
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==12.6.0
roboflow==1.1.34
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
Scrapy==2.9.0
seaborn==0.12.2
secure-smtplib==0.1.1
semver==2.13.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.10.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
sqlparse==0.4.3
stack-data==0.6.0
starlette==0.22.0
stop-words==2018.7.23
streamlit==1.22.0
sympy==1.11.1
tb-nightly==2.14.0a20230511
tenacity==8.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.1.0
terminado==0.17.0
textblob==0.17.1
tf-estimator-nightly==2.14.0.dev2023051108
tf-nightly==2.14.0.dev20230511
tf_nightly_intel==2.14.0.dev20230511
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
tk==0.1.0
tldextract==3.4.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.27.1
tweepy==4.12.1
twilio==7.15.2
twine==4.0.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typer==0.4.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.1
ultralytics==8.1.1
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
virtualenv==20.17.1
w3lib==2.1.1
watchdog==2.3.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
whitenoise==6.3.0
widgetsnbextension==4.0.3
wrapt==1.14.1
yarl==1.8.2
youtube-comment-scraper-python==1.0.0
youtube-dl==2021.12.17
zipp==3.15.0
zope.interface==6.0
